Move sync api files to existing service/apis module
