+++ server/service/src/apis/mod.rs
@@ -1,2 +1,5 @@
 pub mod login_v4;
 pub mod permissions;
+pub mod sync_api_v3;
+pub mod sync_api_v5;
+pub mod sync_api_credentials;
